requirements.txt
fastapi
uvicorn[standard]
rembg
pillow
python-multipart